app/requirements.txt
kafka-python==2.0.2
pymongo==4.5.0
Faker==19.3.0